macOS principal mapping APIs
