+++ docs/api/security/unix/index.md
@@ -33,13 +33,15 @@ echo "uid=$info.uid euid=$info.euid"
 
 ### `user_name uid`
 
-Resolves a Unix user ID in the active VFS target.
+Resolves a Unix user ID in the active VFS target. On macOS, `uid` may also
+be a [`uuid.Uuid`](../../uuid/uuid.md) principal, resolved to a uid first via
+[`security.macos.id_for_uuid`](../macos/index.md#id_for_uuid-uuid).
 
 #### Parameters
 
-| Name  | Type  | Description  |
-| ----- | ----- | ------------ |
-| `uid` | `Int` | Unix user ID |
+| Name  | Type               | Description                |
+| ----- | ------------------ | -------------------------- |
+| `uid` | `Int`\|`uuid.Uuid` | Unix user ID or macOS UUID |
 
 #### Returns
 
@@ -49,7 +51,8 @@ Resolves a Unix user ID in the active VFS target.
 
 - Raises [`sys.NotFoundError`](../../sys/not-found-error.md) when the ID is
   unknown.
-- Raises `UnsupportedError` when the active VFS target is not Unix.
+- Raises `UnsupportedError` when the active VFS target is not Unix, or when
+  a `uuid.Uuid` is passed and the target is not macOS.
 
 ### `user_id name`
 
@@ -67,7 +70,15 @@ Resolves a Unix user name in the active VFS target.
 
 ### `group_name gid`
 
-Resolves a Unix group ID in the active VFS target.
+Resolves a Unix group ID in the active VFS target. On macOS, `gid` may also
+be a [`uuid.Uuid`](../../uuid/uuid.md) principal, resolved to a gid first via
+[`security.macos.id_for_uuid`](../macos/index.md#id_for_uuid-uuid).
+
+#### Parameters
+
+| Name  | Type               | Description                 |
+| ----- | ------------------ | --------------------------- |
+| `gid` | `Int`\|`uuid.Uuid` | Unix group ID or macOS UUID |
 
 #### Returns
 
@@ -77,7 +88,8 @@ Resolves a Unix group ID in the active VFS target.
 
 - Raises [`sys.NotFoundError`](../../sys/not-found-error.md) when the ID is
   unknown.
-- Raises `UnsupportedError` when the active VFS target is not Unix.
+- Raises `UnsupportedError` when the active VFS target is not Unix, or when
+  a `uuid.Uuid` is passed and the target is not macOS.
 
 ### `group_id name`
