+++ dolang-ext-uuid/src/lib.rs
@@ -6,4 +6,4 @@ mod guid;
 mod uuid;
 
 pub use self::guid::{create_guid, value_to_guid};
-pub use self::uuid::{create_uuid, value_to_uuid};
+pub use self::uuid::{cast_uuid, create_uuid, value_to_uuid};
